+++ packages/cli/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
+## Recent Changes
+
 All notable changes to the Zowe CLI package will be documented in this file.
+
+- Enhancement: Added new aliases for zos-files commands in delete, download, and list relating to USS files. You can now interact with `uf` or `uss`.  [#983](https://github.com/zowe/zowe-cli/issues/983)
 
 ## `6.31.0`